requirements.txt
django
django-filter
djangorestframework

django-environ
djangorestframework-simplejwt

Pillow
psycopg2-binary

gunicorn
cryptography

# rabbitmq
pika
